Updated the file ingestion tests for the HDF5 data reader to write temporary files to a proper temp directory.  Removed spurious data sample.
