+++ src/data_readers/unit_test/CMakeLists.txt
@@ -39,6 +39,7 @@ set_full_path(THIS_DIR_MPI_CATCH2_TEST_FILES
   data_reader_HDF5_hrrl_public_api.cpp
   data_reader_HDF5_test.cpp
   data_reader_common_HDF5_test_utils.cpp
+  data_reader_common_catch2.cpp
   data_reader_HDF5_sample_list_test.cpp
   data_reader_synthetic_test_public_api.cpp
   )
