+++ src/data_readers/unit_test/data_reader_common_catch2.hpp
@@ -39,20 +39,7 @@
 namespace pb = ::google::protobuf;
 
 /** create a directory in /tmp; returns the pathname to the directory */
-std::string create_test_directory(std::string base_name)
-{
-  char b[2048];
-  std::stringstream s;
-  s << "/tmp/" << base_name << "_" << getpid();
-  const std::string dir = s.str();
-  lbann::file::make_directory(dir);
-  // test that we can write files
-  sprintf(b, "%s/test", dir.c_str());
-  std::ofstream out(b);
-  REQUIRE(out.good());
-  out.close();
-  return dir;
-}
+std::string create_test_directory(std::string base_name);
 
 /** Instantiates one or more data readers from the input 'prototext' string.
  *  Users should ensure that the appropriate options (if any) are set prior
@@ -66,46 +53,8 @@ instantiate_data_readers(std::string prototext_in,
                          lbann::generic_data_reader*& train_ptr,
                          lbann::generic_data_reader*& validate_ptr,
                          lbann::generic_data_reader*& test_ptr,
-                         lbann::generic_data_reader*& tournament_ptr)
-{
-  lbann_data::LbannPB my_proto;
-  if (!pb::TextFormat::ParseFromString(prototext_in, &my_proto)) {
-    throw "Parsing protobuf failed.";
-  }
+                         lbann::generic_data_reader*& tournament_ptr);
 
-  std::map<lbann::execution_mode, lbann::generic_data_reader*> data_readers;
-  lbann::init_data_readers(&comm_in, my_proto, data_readers);
-
-  // get pointers the the various readers
-  train_ptr = nullptr;
-  validate_ptr = nullptr;
-  test_ptr = nullptr;
-  for (auto t : data_readers) {
-    if (t.second->get_role() == "train") {
-      train_ptr = t.second;
-    }
-    if (t.second->get_role() == "validate") {
-      validate_ptr = t.second;
-    }
-    if (t.second->get_role() == "tournament") {
-      tournament_ptr = t.second;
-    }
-    if (t.second->get_role() == "test") {
-      test_ptr = t.second;
-    }
-  }
-
-  return data_readers;
-}
-
-void write_file(std::string data, std::string dir, std::string fn)
-{
-  std::stringstream s;
-  s << dir << "/" << fn;
-  std::ofstream out(s.str().c_str());
-  REQUIRE(out);
-  out << data;
-  out.close();
-}
+void write_file(std::string data, std::string dir, std::string fn);
 
 #endif //__DATA_READER_TEST_COMMON_HPP__
